修改了打印16进制的输出函数，使用位运算来计算，而不是用除法，对于负数不能求出对应的值

+++ src/test5/include/pmm.h
@@ -61,8 +61,8 @@ void pmm_free_page(uint32 p);
 /*kernel_start是内核加载到内存的起始位置
  *kernel_end 是内核加载到内存的末位置，在链接文件中定义的
  */
-extern uint8 kernel_start[];
-extern uint8 kernel_end[];
+extern uint32 kernel_start[];
+extern uint32 kernel_end[];
 
 
 /*
